+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   based on extractor info received. For extractors using this method they will not need
   to manually register themselves through API to be listed.
   [CATS-1004](https://opensource.ncsa.illinois.edu/jira/browse/CATS-1004)
+- Added support for extractor categories that can be used for retrieving filtered lists of extrators by category.
 
 ### Changed
 - Improved Advanced Search UI to retain search results between navigations.